refactor(gateway): consolidate workspace inventory helpers
Amp-Thread-ID: https://ampcode.com/threads/T-01a037b7-55c9-77ca-b84a-467a6aed97fb
